Modify FreeRTOS and FMC sample code.
FreeRTOS/Blinky: add main_full() demo.
FMC_ExecuteInSRAM: fix IAR and GCC implement.
FMC_MultiBoot: fix GCC implement.
FMC_MultiWordProgram: fix IAR and GCC, FMC code not run in SRAM.
XOMLib: fix IAR implement.
fmc.c: fix FMC_SetSPKey() erase KPROM the second page address.

+++ Library/StdDriver/src/timer.c
@@ -348,7 +348,7 @@ void TIMER_EnableFreqCounter(TIMER_T *timer,
         t = 0UL ;
     }
 
-    if(t != 0UL)
+    if((uint32_t)t != 0UL)
     {
         t->CMP = 0xFFFFFFUL;
         t->EXTCTL = u32EnableInt ? TIMER_EXTCTL_CAPIEN_Msk : 0UL;
